backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
groq==0.18.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic-settings==2.7.1
slowapi==0.1.9
email-validator==2.2.0
python-jose[cryptography]==3.3.0
firebase-admin==6.6.0
